requirements.txt
# Analog Layout Agent - Requirements
# Python >= 3.10

# Core Layout Dependencies (CRITICAL: gdsfactory must be <=7.7.0 for glayout compatibility)
gdsfactory>=6.0.0,<=7.7.0
gdstk>=0.9.0
klayout>=0.29.0
shapely>=2.0.0

# glayout (install separately from source)
# pip install -e ./gLayout

# PydanticAI Framework (replacing OpenAI Agent SDK)
pydantic-ai-slim>=1.47.0
openai>=2.11.0

# OpenAI Agent SDK (deprecated, kept for reference)
# openai-agents>=0.3.0

# MCP Protocol
mcp>=1.0.0
httpx>=0.25.0
httpx-sse>=0.4.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
ipywidgets>=8.0.0

# Async Support
anyio>=4.0.0

# Utilities
loguru>=0.7.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development (optional)
# ipython>=8.0.0
# jupyter>=1.0.0
